UI overflow fix, organize, isolate email functionality
