+++ frontend/src/App.css
@@ -1,11 +1,17 @@
 .App {
   text-align: center;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  background-color: #f1f1f1;
 }
 
 .nav {
   background-color: #f1f1f1c3;
   display: flex;
-
+  padding: 10px 10px;
+  height: "60px";
+  background: #cfdad2;
 }
 
 
@@ -22,7 +28,7 @@
 
 .App-header {
   background-color: #f1f1f1;
-  min-height: 100vh;
+  height: 100%;
   display: flex;
   flex-direction: column;
   align-items: center;
