+++ frontend/src/assets/stockGraphs/AAPL.png
