+++ frontend/src/assets/stockGraphs/AMZN.png
